moved the contents
moved the contents of the file header.js into header-modal.js
