+++ src/index.js
@@ -1,2 +1,3 @@
 import './js/catalog/catalog';
 import './js/header/header';
+import './js/header/header-modal';
